+++ src/test/java/wooteco/subway/acceptance/station/StationAcceptanceTest.java
@@ -2,8 +2,8 @@ package wooteco.subway.acceptance.station;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import wooteco.subway.acceptance.AcceptanceTest;
 import wooteco.subway.service.station.dto.StationResponse;
-import wooteco.subway.AcceptanceTest;
 
 import java.util.List;
